Extract input class to its own file
